new commit

+++ routes/blog.js
@@ -6,12 +6,21 @@ const db = require('../models/index.js');
 // Creating new Blog 
 router.post('/create-blog', async(req, res) => {
     try {
-        const blogData = req.body;
-        const newBlog = await db.blog.create(blogData);
-        // console.log(newBlog);
-        res.status(200).send("Blog created");
+        if(!req.body || !req.body.title || !req.body.description) {
+            return res.status(400).send("Please fill all the fields");
+        } else {
+            const blogData = req.body;
+            if(req.body.title.length > 20) {
+                return res.send("Title exceeds the limit");
+            }
+            const newBlog = await db.blog.create(blogData);
+            console.log(newBlog);
+            return res.status(200).send("Blog created");   
+        }
     } catch (error) {
-        res.status(500).json(error.message);
+        res.status(500).json({
+            msg: error.message,
+        });
     }
 })
 
@@ -24,7 +33,6 @@ router.get('/get-all-blogs', async(req, res) => {
                 model: db.comment
             }]
         })
-        console.log(blogs);
         res.status(200).send(blogs);
     } catch (error) {
         res.status(500).json(error.message);
@@ -35,27 +43,37 @@ router.get('/get-all-blogs', async(req, res) => {
 router.get('/blog/:id', async(req, res) => {
     try {
         const getBlog = await db.blog.findOne({
-            where: {id: req.params.id}
+            where: {id: req.params.id},
+            include: [{
+                model: db.comment
+            }]
         });
-
-        // console.log(getBlog);
-        res.status(200).send(getBlog);
+        // If blog is not found show error else show the blog
+        if(!getBlog) {
+            return res.status(404).send(`Blog with id ${req.params.id} not found`);
+        } else {
+            return res.status(200).send(getBlog);
+        }
     } catch (error) {
-        res.status(500).json(error.message);
+        res.status(404).json(error.message);
     }
 })
 
 // Updating a blog
 router.put('/blog/:id', async(req, res) => {
     try {
+        // get blog from db
         const getBlogToBeUpdated = await db.blog.findOne({
             where: {id: req.params.id}
         });
-        const updatedData = req.body;
-        const updatedBlog = await getBlogToBeUpdated.update({ updatedData });
-
-        // console.log(updatedBlog);
-        res.status(200).send(updatedBlog);
+        // if blog not found then show error else update the blog
+        if(!getBlogToBeUpdated) {
+            return res.status(404).send("This blog does not exist");
+        } else {
+            const updatedBlog = await getBlogToBeUpdated.update( { description: "Some edited content"});
+            // console.log(updatedBlog);
+            res.status(200).send("Blog updated successfully");
+        }
     } catch (error) {
         res.status(500).json(error.message);
     }
@@ -64,16 +82,22 @@ router.put('/blog/:id', async(req, res) => {
 // Deleting a blog
 router.delete('/blog/:id', async(req, res) => {
     try {
-        const removeBlog = await db.blog.destroy({
-            where: { id:req.params.id }
+        // find the blog
+        const getBlogToBeDeleted = await db.blog.findOne({
+            where: {id: req.params.id}
         });
-        
-        // console.log("deleted");
-        res.send("Blog deleted");
+        // If blog is not found then show error else delete the blog
+        if(!getBlogToBeDeleted) {
+            return res.status(404).send(`Blog with ${req.params.id} does not exist`);
+        } else {
+            const removeBlog = await db.blog.destroy({
+                where: { id:req.params.id }
+            });
+            return res.send(`Blog with id ${req.params.id} is deleted`);
+        }
     } catch (error) {
         res.status(500).json(error.message);
     }
-    
 })
 
 module.exports = router;
